+++ cmd/edit.go
@@ -36,7 +36,6 @@ func editCmd(config *dao.Config, configErr *error) *cobra.Command {
 	}
 
 	cmd.AddCommand(
-		editDir(config, configErr),
 		editTask(config, configErr),
 		editProject(config, configErr),
 	)