+++ cmd/exec.go
@@ -22,10 +22,10 @@ Single quote your command if you don't want the file globbing and environments v
 before the command gets executed in each directory.`,
 
 		Example: `  # List files in all projects
-  mani exec ls --all-projects
+  mani exec ls --all
 
   # List all git files that have markdown suffix
-  mani exec 'git ls-files | grep -e ".md"' --all-projects`,
+  mani exec 'git ls-files | grep -e ".md"' --all`,
 		Args: cobra.MinimumNArgs(1),
 		Run: func(cmd *cobra.Command, args []string) {
 			core.CheckIfError(*configErr)
@@ -48,8 +48,7 @@ before the command gets executed in each directory.`,
 
 	cmd.Flags().BoolVarP(&runFlags.Cwd, "cwd", "k", false, "current working directory")
 
-	cmd.Flags().BoolVar(&runFlags.AllProjects, "all-projects", false, "target all projects")
-	cmd.Flags().BoolVar(&runFlags.AllDirs, "all-dirs", false, "target all dirs")
+	cmd.Flags().BoolVarP(&runFlags.AllProjects, "all", "a", false, "target all projects")
 
 	cmd.Flags().StringSliceVarP(&runFlags.Projects, "projects", "p", []string{}, "target projects by their name")
 	err = cmd.RegisterFlagCompletionFunc("projects", func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
@@ -62,27 +61,13 @@ before the command gets executed in each directory.`,
 	})
 	core.CheckIfError(err)
 
-	cmd.Flags().StringSliceVarP(&runFlags.Dirs, "dirs", "d", []string{}, "target directories by their name")
-	err = cmd.RegisterFlagCompletionFunc("dirs", func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
-		if *configErr != nil {
-			return []string{}, cobra.ShellCompDirectiveDefault
-		}
-
-		dirs := config.GetDirNames()
-		return dirs, cobra.ShellCompDirectiveDefault
-	})
-	core.CheckIfError(err)
-
 	cmd.Flags().StringSliceVarP(&runFlags.Paths, "paths", "g", []string{}, "target directories by their path")
 	err = cmd.RegisterFlagCompletionFunc("paths", func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
 		if *configErr != nil {
 			return []string{}, cobra.ShellCompDirectiveDefault
 		}
 
-		projectPaths := config.GetProjectPaths()
-		dirPaths := config.GetDirPaths()
-
-		options := append(projectPaths, dirPaths...)
+		options := config.GetProjectPaths()
 
 		return options, cobra.ShellCompDirectiveDefault
 	})
@@ -107,27 +92,12 @@ func execute(
 	config *dao.Config,
 	runFlags core.RunFlags,
 ) {
-	projectEntities, dirEntities := config.GetEntities(runFlags)
+	projects := config.FilterProjects(runFlags.Cwd, runFlags.AllProjects, runFlags.Paths, runFlags.Projects, runFlags.Tags)
 
-	if len(projectEntities) == 0 && len(dirEntities) == 0 {
+	if len(projects) == 0 {
 		fmt.Println("No targets")
 	} else {
 		cmd := strings.Join(args[0:], " ")
-		if len(projectEntities) > 0 {
-			entityList := dao.EntityList{
-				Type:     "Project",
-				Entities: projectEntities,
-			}
-
-			dao.RunExec(cmd, entityList, config, &runFlags)
-		}
-
-		if len(dirEntities) > 0 {
-			entityList := dao.EntityList{
-				Type:     "Directory",
-				Entities: dirEntities,
-			}
-			dao.RunExec(cmd, entityList, config, &runFlags)
-		}
+		dao.RunExec(cmd, projects, config, &runFlags)
 	}
 }