+++ cmd/sync.go
@@ -14,24 +14,23 @@ func syncCmd(config *dao.Config, configErr *error) *cobra.Command {
 		Use:   "sync",
 		Short: "Clone repositories and add them to gitignore",
 		Long: `Clone repositories and add them to gitignore.
-In-case you need to enter credentials before cloning, run the command with the parallel flag.`,
+In-case you need to enter credentials before cloning, run the command without the parallel flag.`,
 		Example: `  # Clone repositories one at a time
   mani sync
 
-  # Clone repositories one by one
-  mani sync --parallel=false`,
+  # Clone repositories in parallell
+  mani sync --parallel`,
 		Run: func(cmd *cobra.Command, args []string) {
 			core.CheckIfError(*configErr)
 			runSync(config, syncFlags)
 		},
 	}
 
-	cmd.Flags().BoolVarP(&syncFlags.Parallel, "parallel", "p", true, "Clone projects in parallel")
+	cmd.Flags().BoolVarP(&syncFlags.Parallel, "parallel", "p", false, "Clone projects in parallel")
 
 	return &cmd
 }
 
 func runSync(config *dao.Config, syncFlags core.SyncFlags) {
-	config.SyncDirs(config.Dir, syncFlags.Parallel)
 	config.SyncProjects(config.Dir, syncFlags.Parallel)
 }