+++ core/dao/exec.go
@@ -14,21 +14,21 @@ import (
 
 func RunExec(
 	cmd string,
-	entityList EntityList,
+	projects []Project,
 	config *Config,
 	runFlags *core.RunFlags,
 ) {
 	switch runFlags.Output {
 	case "table", "html", "markdown" :
-		tableExec(cmd, entityList, config, runFlags)
+		tableExec(cmd, projects, config, runFlags)
 	default: // text
-		textExec(cmd, entityList, config, runFlags)
+		textExec(cmd, projects, config, runFlags)
 	}
 }
 
 func tableExec(
 	cmd string,
-	entityList EntityList,
+	projects []Project,
 	config *Config,
 	runFlags *core.RunFlags,
 ) {
@@ -43,13 +43,13 @@ func tableExec(
 	/**
 	** Headers
 	**/
-	data.Headers = append(data.Headers, entityList.Type)
+	data.Headers = append(data.Headers, "Project")
 
 	// Append Command name if set
 	data.Headers = append(data.Headers, "Output")
 
-	for _, entity := range entityList.Entities {
-		data.Rows = append(data.Rows, table.Row{entity.Name})
+	for _, project := range projects {
+		data.Rows = append(data.Rows, table.Row{project.Name})
 	}
 
 	/**
@@ -57,15 +57,15 @@ func tableExec(
 	**/
 	var wg sync.WaitGroup
 
-	for i, entity := range entityList.Entities {
+	for i, project := range projects {
 		wg.Add(1)
 
 		if runFlags.Parallel {
 			spinner.Message(" Running")
-			go tableWork(config, &data, cmd, entity, runFlags.DryRun, i, &wg)
+			go tableWork(config, &data, cmd, project, runFlags.DryRun, i, &wg)
 		} else {
-			spinner.Message(fmt.Sprintf(" %v", entity.Name))
-			tableWork(config, &data, cmd, entity, runFlags.DryRun, i, &wg)
+			spinner.Message(fmt.Sprintf(" %v", project.Name))
+			tableWork(config, &data, cmd, project, runFlags.DryRun, i, &wg)
 		}
 	}
 
@@ -84,7 +84,7 @@ func tableWork(
 	config *Config,
 	data *core.TableOutput,
 	cmd string,
-	entity Entity,
+	project Project,
 	dryRunFlag bool,
 	i int,
 	wg *sync.WaitGroup,
@@ -92,26 +92,26 @@ func tableWork(
 	defer wg.Done()
 
 	var output string
-	output, _ = RunTable(*config, cmd, []string{}, config.Shell, entity, dryRunFlag)
+	output, _ = RunTable(*config, cmd, []string{}, config.Shell, project, dryRunFlag)
 	data.Rows[i] = append(data.Rows[i], strings.TrimSuffix(output, "\n"))
 }
 
 func textExec(
 	cmd string,
-	entityList EntityList,
+	projects []Project,
 	config *Config,
 	runFlags *core.RunFlags,
 ) {
 	var wg sync.WaitGroup
 
-	for i, entity := range entityList.Entities {
+	for i, project := range projects {
 		colorIndex := core.COLOR_INDEX[i % len(core.COLOR_INDEX)]
 
 		wg.Add(1)
 		if runFlags.Parallel {
-			go textWork(uint8(colorIndex), config, cmd, entity, runFlags.DryRun, &wg)
+			go textWork(uint8(colorIndex), config, cmd, project, runFlags.DryRun, &wg)
 		} else {
-			textWork(uint8(colorIndex), config, cmd, entity, runFlags.DryRun, &wg)
+			textWork(uint8(colorIndex), config, cmd, project, runFlags.DryRun, &wg)
 		}
 	}
 
@@ -122,13 +122,13 @@ func textWork(
 	colorIndex uint8,
 	config *Config,
 	cmd string,
-	entity Entity,
+	project Project,
 	dryRunFlag bool,
 	wg *sync.WaitGroup,
 ) {
 	defer wg.Done()
 
-	header := fmt.Sprintf("[%s]", color.Index(colorIndex, entity.Name))
+	header := fmt.Sprintf("[%s]", color.Index(colorIndex, project.Name))
 
 	width, _, err := term.GetSize(0)
 	core.CheckIfError(err)
@@ -136,6 +136,6 @@ func textWork(
 	header = fmt.Sprintf("\n%s %s\n", header, strings.Repeat("*", width - headerLength - 1))
 	fmt.Println(header)
 
-	err = RunText(cmd, []string{}, *config, config.Shell, entity, dryRunFlag)
+	err = RunText(cmd, []string{}, *config, config.Shell, project, dryRunFlag)
 	core.CheckIfError(err)
 }