+++ _site/docs/changelog.md
@@ -3,6 +3,8 @@
 ## [UNRELEASED]
 
 - Enable `mani edit` to run even if config file is malformed (wrong YAML syntax)
+- Remove dirs entity
+- [BREAKING CHANGE]: Move tree feature to list projects as a flag instead of it being a special sub-command
 
 ## v0.11.1